chore: move ./exp/cloud into ./cloud
