+++ cloud/services/scalesets/mock_scalesets/scalesets_mock.go
